fix(rust): Fix cross-platform type compatibility for error codes
Problem Analysis:
- Windows builds were failing with type mismatch errors in error.rs and provider.rs
- Root cause: CcapErrorCode constants have platform-dependent types
  - Linux/macOS: c_uint (u32) via GCC/Clang
  - Windows: c_int (i32) via MSVC
- Original code assumed u32 and cast i32 to u32, causing type errors on Windows

Error Details:
- error.rs:97: expected u32, found i32 for CcapErrorCode constants
- provider.rs:353: error callback expected u32 but received i32

Changes:
1. error.rs: Convert i32 to CcapErrorCode type before matching
   - Use 'let code_u = code as CcapErrorCode' to handle platform differences
   - This ensures constants match regardless of underlying type

2. provider.rs: Change error callback signature from Fn(u32, &str) to Fn(i32, &str)
   - Use i32 as the common error code type across all platforms
   - Cast sys::CcapErrorCode to i32 when calling the callback

Verification:
- Linux: ✅ cargo build --features static-link
- Linux: ✅ cargo clippy --all-targets --features static-link -- -D warnings
- Linux: ✅ cargo build --no-default-features --features build-source
- Expected to fix Windows CI failures

This ensures the Rust bindings work correctly on all platforms despite
C enum type differences between compilers.

+++ bindings/rust/src/provider.rs
@@ -330,7 +330,7 @@ impl Provider {
     /// Set error callback for camera errors
     pub fn set_error_callback<F>(callback: F)
     where
-        F: Fn(u32, &str) + Send + Sync + 'static,
+        F: Fn(i32, &str) + Send + Sync + 'static,
     {
         use std::os::raw::c_char;
         use std::sync::{Arc, Mutex};
@@ -346,11 +346,11 @@ impl Provider {
                 return;
             }
 
-            let callback = &*(user_data as *const Arc<Mutex<dyn Fn(u32, &str) + Send + Sync>>);
+            let callback = &*(user_data as *const Arc<Mutex<dyn Fn(i32, &str) + Send + Sync>>);
             let desc_cstr = std::ffi::CStr::from_ptr(description);
             if let Ok(desc_str) = desc_cstr.to_str() {
                 if let Ok(cb) = callback.lock() {
-                    cb(error_code, desc_str);
+                    cb(error_code as i32, desc_str);
                 }
             }
         }
